refactor low latency
