--- tools/camerastream/.gitignore
@@ -1,1 +1,0 @@
-compressed_low_latency
