+++ tools/camerastream/low_latency/.gitignore
@@ -1,0 +1,1 @@
+low_latency
